Change file modes to octal notation.
Noticed that when nodeup was running it was comparing the octal returned
from the operating system with the non-octal prior to this change and
causing the files to be rewritten with the wrong owners (in the case of
aws-iam-authenticator). Using the octal representation in the string
makes the comparison work and the correct files are written.